fix #107: edit service structure
